add upsert behavior to avoid concurrence errors
