--- app/jobs/outgoing_webhooks/france_travail/update_participation_job.rb
@@ -1,13 +1,0 @@
-module OutgoingWebhooks
-  module FranceTravail
-    class UpdateParticipationJob < BaseJob
-      def perform(participation_id:, timestamp:)
-        participation = Participation.find_by(id: participation_id)
-
-        return unless participation
-
-        call_service!(FranceTravailApi::UpdateParticipation, participation:, timestamp:)
-      end
-    end
-  end
-end
